Rename javascript tools from .js to .mjs
