+++ Dockerfile
@@ -14,9 +14,9 @@ ENV ARCHIVE="foundryvtt-${FOUNDRY_VERSION}.zip"
 
 WORKDIR /root
 COPY \
-  src/authenticate.js \
-  src/get_release_url.js \
-  src/logging.js \
+  src/authenticate.mjs \
+  src/get_release_url.mjs \
+  src/logging.mjs \
   src/package.json \
   ./
 # .placeholder file to mitigate https://github.com/moby/moby/issues/37965
@@ -24,8 +24,8 @@ RUN mkdir dist && touch dist/.placeholder
 RUN \
   if [ -n "${FOUNDRY_USERNAME}" ] && [ -n "${FOUNDRY_PASSWORD}" ]; then \
   npm install && \
-  ./authenticate.js "${FOUNDRY_USERNAME}" "${FOUNDRY_PASSWORD}" cookiejar.json && \
-  s3_url=$(./get_release_url.js cookiejar.json "${FOUNDRY_VERSION}") && \
+  ./authenticate.mjs "${FOUNDRY_USERNAME}" "${FOUNDRY_PASSWORD}" cookiejar.json && \
+  s3_url=$(./get_release_url.mjs cookiejar.json "${FOUNDRY_VERSION}") && \
   wget -O ${ARCHIVE} "${s3_url}" && \
   unzip -d dist ${ARCHIVE} 'resources/*'; \
   elif [ -n "${FOUNDRY_RELEASE_URL}" ]; then \
@@ -51,18 +51,18 @@ WORKDIR ${FOUNDRY_HOME}
 
 COPY --from=optional-release-stage /root/dist/ .
 COPY \
-  src/authenticate.js \
+  src/authenticate.mjs \
   src/check_health.sh \
   src/entrypoint.sh \
-  src/get_license.js \
-  src/get_release_url.js \
+  src/get_license.mjs \
+  src/get_release_url.mjs \
   src/launcher.sh \
-  src/logging.js \
+  src/logging.mjs \
   src/logging.sh \
   src/package.json \
   src/patch_lang.mjs \
-  src/set_options.js \
-  src/set_password.js \
+  src/set_options.mjs \
+  src/set_password.mjs \
   ./
 RUN addgroup --system --gid ${FOUNDRY_UID} foundry \
   && adduser --system --uid ${FOUNDRY_UID} --ingroup foundry foundry \
@@ -84,6 +84,6 @@ EXPOSE 30000/TCP
 # EXPOSE 49152-65535/UDP
 
 ENTRYPOINT ["./entrypoint.sh"]
-CMD ["resources/app/main.js", "--port=30000", "--headless", "--noupdate",\
+CMD ["resources/app/main.mjs", "--port=30000", "--headless", "--noupdate",\
   "--dataPath=/data"]
 HEALTHCHECK --start-period=3m --interval=30s --timeout=5s CMD ./check_health.sh
